sonarLint 기반 source refactoring

+++ src/main/java/com/tsp/new_tsp_admin/api/portfolio/service/AdminPortfolioJpaService.java
@@ -18,7 +18,7 @@ public interface AdminPortfolioJpaService {
      * </pre>
      *
      */
-    Long findPortfoliosCount(Map<String, Object> portfolioMap);
+    int findPortfoliosCount(Map<String, Object> portfolioMap);
 
     /**
      * <pre>
